fix Windows

+++ exporters/prometheus/test/prometheus_test_helper.h
@@ -54,7 +54,7 @@ inline metric_sdk::ResourceMetrics CreateHistogramPointData()
   histogram_point_data2.boundaries_ = {10.0, 20.0, 30.0};
   histogram_point_data2.count_      = 3;
   histogram_point_data2.counts_     = {200, 300, 400, 500};
-  histogram_point_data2.sum_        = 900l;
+  histogram_point_data2.sum_        = (int64_t)900;
   metric_sdk::ResourceMetrics data;
   auto resource = opentelemetry::sdk::resource::Resource::Create(
       opentelemetry::sdk::resource::ResourceAttributes{});
@@ -90,7 +90,7 @@ inline metric_sdk::ResourceMetrics CreateLastValuePointData()
   last_value_point_data.is_lastvalue_valid_ = true;
   last_value_point_data.sample_ts_          = opentelemetry::common::SystemTimestamp{};
   metric_sdk::LastValuePointData last_value_point_data2{};
-  last_value_point_data2.value_              = 20l;
+  last_value_point_data2.value_              = (int64_t)20;
   last_value_point_data2.is_lastvalue_valid_ = true;
   last_value_point_data2.sample_ts_          = opentelemetry::common::SystemTimestamp{};
   metric_sdk::MetricData metric_data{
