+++ sdk/src/metrics/aggregation/histogram_aggregation.cc
@@ -33,7 +33,7 @@ LongHistogramAggregation::LongHistogramAggregation(
     record_min_max_ = aggregation_config->record_min_max_;
   }
   point_data_.counts_         = std::vector<uint64_t>(point_data_.boundaries_.size() + 1, 0);
-  point_data_.sum_            = 0l;
+  point_data_.sum_            = (int64_t)0;
   point_data_.count_          = 0;
   point_data_.record_min_max_ = record_min_max_;
   point_data_.min_            = std::numeric_limits<int64_t>::max();
