+++ sdk/test/metrics/observable_registry_test.cc
@@ -21,15 +21,15 @@ public:
     {
       std::get<observer_result.Observe(20l, {{"RequestType", "GET"}});
       observer_result.Observe(10l, {{"RequestType", "PUT"}});
-      number_of_get1 += 20l;
-      number_of_put1 += 10l;
+      number_of_get1 += (int64_t)20;
+      number_of_put1 += (int64_t)10;
     }
     else if (fetch_count1 == 2)
     {
       observer_result.Observe(40l, {{"RequestType", "GET"}});
       observer_result.Observe(20l, {{"RequestType", "PUT"}});
-      number_of_get1 += 40l;
-      number_of_put1 += 20l;
+      number_of_get1 += (int64_t)40;
+      number_of_put1 += (int64_t)20;
     }
   }
 
@@ -41,15 +41,15 @@ public:
     {
       observer_result.Observe(20l, {{"RequestType", "GET"}});
       observer_result.Observe(10l, {{"RequestType", "PUT"}});
-      number_of_get2 += 20l;
-      number_of_put2 += 10l;
+      number_of_get2 += (int64_t)20;
+      number_of_put2 += (int64_t)10;
     }
     else if (fetch_count2 == 2)
     {
       observer_result.Observe(40l, {{"RequestType", "GET"}});
       observer_result.Observe(20l, {{"RequestType", "PUT"}});
-      number_of_get2 += 40l;
-      number_of_put2 += 20l;
+      number_of_get2 += (int64_t)40;
+      number_of_put2 += (int64_t)20;
     }
   }
